added queue for noteoffs

+++ Scales.h
@@ -1,18 +1,16 @@
 namespace SCALE 
 {
-    //how to add a scale to the application - enter instructions here!
-    
-    const int chromatic[12] = {0,1,2,3,4,5,6,7,8,9,10,11};
-    const int major[12] = {0,2,4,5,7,9,11,12,14,16,17,19};
-    const int majorPentatonic[12] = {0,2,4,7,9,12,14,16,19,21,24,26};
-    const int minor[12] = {0,2,3,5,7,8,10,12,14,15,17,19};
-    const int minorPentatonic[12] = {0,3,5,7,10,12,15,17,19,22,24,26};
-    const int melodicMinor[12] = {0,2,3,5,7,9,11,12,14,15,17,19};
-    const int harmonicMinor[12] = {0,2,3,5,7,8,11,12,14,15,17,19};
-    const int blues[12] = {0,3,5,6,7,10,12,15,17,18,19,22};
-    const int rockNRoll[12] = {0,3,4,5,7,9,10,12,15,16,17,19};
-    const int hungarianGypsy[12] = {0,2,3,6,7,8,11,12,14,15,18,19};
-    const int ragaMadhuri[12] = {0,4,5,7,9,10,11,12,16,17,19,21};
-    const int alphaAdam[12] = {0,2,4,6,7,9,11,13,14,16,18,20}; //???
+    const uint8_t chromatic[12] = {0,1,2,3,4,5,6,7,8,9,10,11};
+    const uint8_t major[12] = {0,2,4,5,7,9,11,12,14,16,17,19};
+    const uint8_t majorPentatonic[12] = {0,2,4,7,9,12,14,16,19,21,24,26};
+    const uint8_t minor[12] = {0,2,3,5,7,8,10,12,14,15,17,19};
+    const uint8_t minorPentatonic[12] = {0,3,5,7,10,12,15,17,19,22,24,26};
+    const uint8_t melodicMinor[12] = {0,2,3,5,7,9,11,12,14,15,17,19};
+    const uint8_t harmonicMinor[12] = {0,2,3,5,7,8,11,12,14,15,17,19};
+    const uint8_t blues[12] = {0,3,5,6,7,10,12,15,17,18,19,22};
+    const uint8_t rockNRoll[12] = {0,3,4,5,7,9,10,12,15,16,17,19};
+    const uint8_t hungarianGypsy[12] = {0,2,3,6,7,8,11,12,14,15,18,19};
+    const uint8_t ragaMadhuri[12] = {0,4,5,7,9,10,11,12,16,17,19,21};
+    const uint8_t alphaAdam[12] = {0,2,4,6,7,9,11,13,14,16,18,20}; //???
     
 }
